lock down delete item permissions

+++ backend/src/resolvers/Mutation.js
@@ -44,7 +44,14 @@ const Mutations = {
 
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
-    const item = await ctx.db.query.item({ where }, `{id title}`)
+    const item = await ctx.db.query.item({ where }, '{id title user { id }}')
+    const ownsItem = Object.is(item.user.id, ctx.request.userId)
+    const hasPermission = ctx.request.user.permissions.some(permision =>
+      ['ADMIN', 'ITEMDELETE'].includes(permision),
+    )
+    if (!ownsItem && !hasPermission) {
+      throw new Error('Nie masz uprawnień!')
+    }
     return ctx.db.mutation.deleteItem({ where }, info)
   },
